+++ lib/Perl/Tidy/Tokenizer.pm
@@ -23,7 +23,7 @@ use strict;
 use warnings;
 use English qw( -no_match_vars );
 
-our $VERSION = '20230309.01';
+our $VERSION = '20230309.02';
 
 use Perl::Tidy::LineBuffer;
 use Carp;